feat: refactor member-related classes to use new DTO package structure
