+++ src/main/java/org/mandarin/booking/adapter/webapi/AuthController.java
@@ -2,9 +2,9 @@ package org.mandarin.booking.adapter.webapi;
 
 import jakarta.validation.Valid;
 import org.mandarin.booking.app.port.AuthUseCase;
-import org.mandarin.booking.domain.member.AuthRequest;
-import org.mandarin.booking.domain.member.ReissueRequest;
-import org.mandarin.booking.domain.member.TokenHolder;
+import org.mandarin.booking.adapter.webapi.dto.AuthRequest;
+import org.mandarin.booking.adapter.webapi.dto.ReissueRequest;
+import org.mandarin.booking.adapter.webapi.dto.TokenHolder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
