+++ src/main/java/org/mandarin/booking/adapter/webapi/dto/MemberRegisterRequest.java
@@ -1,4 +1,4 @@
-package org.mandarin.booking.domain.member;
+package org.mandarin.booking.adapter.webapi.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
