+++ src/main/java/org/mandarin/booking/app/AuthService.java
@@ -2,7 +2,7 @@ package org.mandarin.booking.app;
 
 import lombok.RequiredArgsConstructor;
 import org.mandarin.booking.adapter.persist.MemberQueryRepository;
-import org.mandarin.booking.domain.member.TokenHolder;
+import org.mandarin.booking.adapter.webapi.dto.TokenHolder;
 import org.mandarin.booking.app.port.AuthUseCase;
 import org.mandarin.booking.domain.member.AuthException;
 import org.mandarin.booking.domain.member.Member;
