+++ src/main/java/org/mandarin/booking/app/MemberService.java
@@ -1,11 +1,12 @@
 package org.mandarin.booking.app;
 
 import lombok.RequiredArgsConstructor;
-import org.mandarin.booking.domain.member.MemberRegisterRequest;
-import org.mandarin.booking.domain.member.MemberRegisterResponse;
+import org.mandarin.booking.adapter.webapi.dto.MemberRegisterRequest;
+import org.mandarin.booking.adapter.webapi.dto.MemberRegisterResponse;
 import org.mandarin.booking.app.port.MemberRegisterer;
 import org.mandarin.booking.adapter.persist.MemberCommandRepository;
 import org.mandarin.booking.domain.member.Member;
+import org.mandarin.booking.domain.member.MemberCreateCommand;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -20,8 +21,10 @@ public class MemberService implements MemberRegisterer {
         validator.checkDuplicateUserId(request.userId());
         validator.checkDuplicateEmail(request.email());
 
-        var newMember = Member.create(request, securePasswordEncoder);
-        var savedMember = command.insert(newMember);
+        var createCommand
+                = new MemberCreateCommand(request.nickName(), request.userId(), request.password(), request.email());
+        var newMember = Member.create(createCommand, securePasswordEncoder);
+        var savedMember = this.command.insert(newMember);
         return MemberRegisterResponse.from(savedMember);
     }
 }
