+++ src/main/java/org/mandarin/booking/app/port/AuthUseCase.java
@@ -1,6 +1,6 @@
 package org.mandarin.booking.app.port;
 
-import org.mandarin.booking.domain.member.TokenHolder;
+import org.mandarin.booking.adapter.webapi.dto.TokenHolder;
 
 public interface AuthUseCase {
     TokenHolder login(String userId, String password);
