+++ src/main/java/org/mandarin/booking/domain/member/MemberCreateCommand.java
@@ -1,0 +1,4 @@
+package org.mandarin.booking.domain.member;
+
+public record MemberCreateCommand(String nickName, String userId, String password, String email){
+}
