+++ src/test/java/org/mandarin/booking/webapi/member/POST_specs.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.mandarin.booking.IntegrationTest;
-import org.mandarin.booking.domain.member.MemberRegisterRequest;
-import org.mandarin.booking.domain.member.MemberRegisterResponse;
+import org.mandarin.booking.adapter.webapi.dto.MemberRegisterRequest;
+import org.mandarin.booking.adapter.webapi.dto.MemberRegisterResponse;
 import org.mandarin.booking.app.SecurePasswordEncoder;
 import org.mandarin.booking.fixture.MemberFixture.NicknameGenerator;
 import org.mandarin.booking.fixture.MemberFixture.PasswordGenerator;
